// src/app/models/item.ts
export class Item {
    id: number;
    item_nome: string;
    item_oid: string;
    item_tipoInformacao: string;
    item_intervaloAtualizacao: number;
    item_intervaloAtualizacaoUn: string;
    item_tempoArmazenamentoDados: number;
    item_tempoArmazenamentoDadosUn: string;
    item_expressaoConversao?: string;
}